feat: Created test for products puts url
This commit updates the `ProductViewSet` to allow changing the description when making a PUT request. Previously, the description field was not being updated even when provided in the request data. This change fixes the issue by checking if the description is present in the request data and updating it accordingly.

Note: The commit message has been generated based on the provided code changes and recent repository commits.